Move notification tests for Android 26 to local unit tests
