+++ library/src/test/java/io/ona/kujaku/notifications/BaseNotificationTest.java
@@ -1,5 +1,6 @@
 package io.ona.kujaku.notifications;
 
+import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
@@ -8,6 +9,7 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 
+import org.junit.After;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
@@ -17,7 +19,7 @@ import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowPendingIntent;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
+import java.util.ArrayList;
 
 import io.ona.kujaku.BaseTest;
 import io.ona.kujaku.BuildConfig;
@@ -30,10 +32,11 @@ import static org.junit.Assert.assertEquals;
  * Created by Ephraim Kigamba - ekigamba@ona.io on 16/01/2018.
  */
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, manifest = Config.NONE, shadows = {ShadowNotificationChannel.class})
+@Config(constants = BuildConfig.class, manifest = Config.NONE, sdk = Build.VERSION_CODES.O)
 public abstract class BaseNotificationTest extends BaseTest {
 
     protected Context context;
+    protected ArrayList<String> channelIdsAdded = new ArrayList<>();
 
     @Before
     public void setup() {
@@ -90,6 +93,41 @@ public abstract class BaseNotificationTest extends BaseTest {
 
     protected void setSDKToAndroidOreo() throws NoSuchFieldException, IllegalAccessException {
         setFinalStatic(Build.VERSION.class.getField("SDK_INT"), Build.VERSION_CODES.O);
+    }
+
+    @After
+    public void tearDown() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+
+            for (String channelId : channelIdsAdded) {
+                notificationManager.deleteNotificationChannel(channelId);
+            }
+
+            channelIdsAdded.clear();
+        }
+    }
+
+    protected void insertValueInPrivateField(Class classWithField, Object instance, String fieldName, Object newValue) throws IllegalAccessException, NoSuchFieldException {
+        Field instanceField = classWithField.getDeclaredField(fieldName);
+        if (!instanceField.isAccessible()) {
+            instanceField.setAccessible(true);
+        }
+
+        instanceField.set(instance, newValue);
+    }
+
+    protected void insertValueInPrivateStaticField(Class classWithField, String fieldName, Object newValue) throws NoSuchFieldException, IllegalAccessException {
+        insertValueInPrivateField(classWithField, null, fieldName, newValue);
+    }
+
+    protected Object getValueInPrivateField(Class classWithField, Object instance, String fieldName) throws IllegalAccessException, NoSuchFieldException {
+        Field instanceField = classWithField.getDeclaredField(fieldName);
+        if (!instanceField.isAccessible()) {
+            instanceField.setAccessible(true);
+        }
+
+        return instanceField.get(instance);
     }
 
 }
